requirements.txt
apipkg==1.5
asgiref==3.3.4
astroid==2.5.2
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
colorclass==2.2.0
coverage==5.5
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.2
django-filter==2.4.0
djangorestframework==3.12.4
docopt==0.6.2
execnet==1.8.0
Faker==8.1.0
flake8==3.9.0
gunicorn==20.1.0
idna==3.1
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==20.9
Pillow==8.2.0
pip-upgrade==0.0.6
pip-upgrader==1.4.15
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.7.4
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.1.0
pytest-flake8==1.0.7
pytest-forked==1.3.0
pytest-isort==2.0.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
terminaltables==3.1.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
